thumbnail-generation/requirements.txt
absl-py==2.0.0
addict==2.4.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
decorator==5.1.1
docstring-parser==0.15
einops==0.7.0
executing==2.0.0
fairscale==0.4.13
filelock==3.12.4
fonttools==4.43.0
fsspec==2023.9.2
ftfy==6.1.1
future==0.18.3
google-auth==2.23.2
google-auth-oauthlib==1.1.0
grpcio==1.54.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.5
imgaug==0.4.0
importlib-metadata==6.8.0
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonref==1.1.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.8.1.78
opencv-python-headless==4.5.5.64
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
safetensors==0.3.3
scenedetect==0.6.2
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.1
sievedata @ https://mango.sievedata.com/v1/client_package/sievedata-0.0.1.1.2-py3-none-any.whl#sha256=96dd87ea688763db844ea5a362979792974a237179142abadf0c63c60a6220d7
six==1.16.0
sklearn==0.0.post9
sseclient==0.0.27
stack-data==0.6.3
sympy==1.12
tb-nightly==2.15.0a20231004
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.14.0
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.64.1
traitlets==5.11.2
transformers==4.34.0
typeguard==4.1.5
typer==0.9.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uuid==1.30
wcwidth==0.2.8
Werkzeug==3.0.0
widgetsnbextension==4.0.9
yapf==0.40.2
zipp==3.17.0